@@ -22,7 +22,7 @@ func dataSourceConfig() *schema.Resource {
 				Computed: true,
 			},
 		},
-        Timeouts: &schema.ResourceTimeout{
+		Timeouts: &schema.ResourceTimeout{
 			Read:    schema.DefaultTimeout(10 * time.Minute),
 			Default: schema.DefaultTimeout(10 * time.Minute),
 		},
@@ -39,10 +39,13 @@ func dataSourceConfigRead(ctx context.Context, d *schema.ResourceData, m interfa
 		return diag.FromErr(err)
 	}
 
-	if value != nil {
-		if err := d.Set("value", *value); err != nil {
+	switch value := value.(type) {
+	case string:
+		if err := d.Set("value", value); err != nil {
 			return diag.FromErr(err)
 		}
+	default:
+		return diag.Errorf("Configuration at '%s' is not a string: %s.", key, value)
 	}
 
 	d.SetId(strconv.FormatInt(time.Now().Unix(), 10))
