@@ -36,7 +36,7 @@ func resourceConfig() *schema.Resource {
 				Required:    true,
 			},
 		},
-        Timeouts: &schema.ResourceTimeout{
+		Timeouts: &schema.ResourceTimeout{
 			Create:  schema.DefaultTimeout(10 * time.Minute),
 			Read:    schema.DefaultTimeout(10 * time.Minute),
 			Update:  schema.DefaultTimeout(10 * time.Minute),
@@ -60,7 +60,7 @@ func resourceConfigCreate(ctx context.Context, d *schema.ResourceData, m interfa
 	}
 	// Dont care about sub config blocks
 	if val != nil {
-		return diag.Errorf("Configuration '%s' already exists with value '%s' set, try a resource import instead.", key, *val)
+		return diag.Errorf("Configuration '%s' already exists with value '%s' set, try a resource import instead.", key, val)
 	}
 
 	err = c.Config.Set(ctx, key, value)
